Revert PickerMediaMode extraction.
